+++ src/main/java/org/boot/growup/common/enumerate/InquiryCategory.java
@@ -1,4 +1,4 @@
-package org.boot.growup.source.board.persist.entity;
+package org.boot.growup.common.enumerate;
 
 public enum InquiryCategory {
   PAYMENT, ACCOUNT, TECH_SUPPORT, OTHERS
