+++ src/main/java/org/boot/growup/source/board/dto/response/GetInquiryResponseDTO.java
@@ -1,12 +1,9 @@
 package org.boot.growup.source.board.dto.response;
 
-import jakarta.persistence.criteria.CriteriaBuilder.In;
-import java.time.LocalDateTime;
 import lombok.Builder;
 import lombok.Data;
 import org.boot.growup.source.board.persist.entity.Inquiry;
-import org.boot.growup.source.board.persist.entity.InquiryCategory;
-import org.boot.growup.source.board.persist.entity.Notice;
+import org.boot.growup.common.enumerate.InquiryCategory;
 import org.springframework.data.domain.Page;
 
 @Data
@@ -18,8 +15,6 @@ public class GetInquiryResponseDTO {
   private String title;
   private String content;
   private Boolean isAnswered;
-//  private LocalDateTime createdAt;
-  private String createdAt;
   private long customer;
 
   public static GetInquiryResponseDTO from(Inquiry inquiry) {
@@ -29,7 +24,6 @@ public class GetInquiryResponseDTO {
         .title(inquiry.getTitle())
         .content(inquiry.getContent())
         .isAnswered(inquiry.getIsAnswered())
-        .createdAt("수정 필요")
         .customer(inquiry.getCustomer())
         .build();
   }
