+++ src/main/java/org/boot/growup/source/board/persist/InquiryRepository.java
@@ -6,10 +6,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-@Repository
 public interface InquiryRepository extends JpaRepository<Inquiry, Long> {
-
   Page<Inquiry> findByCustomer(long id, Pageable pageable);
-
   Page<Inquiry> findByIsAnswered(Boolean isAnswered, Pageable pageable);
 }
