+++ src/main/java/org/boot/growup/source/board/persist/NoticeRepository.java
@@ -6,7 +6,6 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-@Repository
 public interface NoticeRepository extends JpaRepository<Notice, Long> {
   Page<Notice> findAll(Pageable page);
 }
