+++ src/main/java/org/boot/growup/source/board/persist/ReplyRepository.java
@@ -4,6 +4,5 @@ import org.boot.growup.source.board.persist.entity.Reply;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-@Repository
 public interface ReplyRepository extends JpaRepository<Reply, Long> {
 }
